Move the hidadpi backend to its own module
This lets us avoid the special-casing in `lib.rs` and we can export whichever
type exists for our backend.

This is hoever not great as the iteration still knows a bit too much, plus we
have no way to statically verify that the types have the same API.
